fix: 캐릭터 수정 로직 리팩토링

+++ src/services/characters.service.ts
@@ -2,7 +2,12 @@ import { ForbiddenException, Injectable, NotFoundException } from '@nestjs/commo
 import { Prisma } from '@prisma/client';
 import { randomUUID } from 'crypto';
 import { Character, CharacterSnapshot } from 'src/interfaces/characters.interface';
+import { Experience } from 'src/interfaces/experiences.interface';
 import { Member } from 'src/interfaces/member.interface';
+import { Personality } from 'src/interfaces/personalities.interface';
+import { Position } from 'src/interfaces/positions.interface';
+import { Skill } from 'src/interfaces/skills.interface';
+import { Source } from 'src/interfaces/source.interface';
 import { DateTimeUtil } from 'src/util/date-time.util';
 import { PaginationUtil } from 'src/util/pagination.util';
 import { ExperiencesService } from './experiences.service';
@@ -171,6 +176,7 @@ export class CharactersService {
               select: { id: true, keyword: true },
             },
           },
+          where: { deleted_at: null },
         },
         _count: { select: { rooms: true } },
       },
@@ -470,7 +476,9 @@ export class CharactersService {
   /**
    * 캐릭터를 수정한다.
    */
-  async update(memberId: Member['id'], id: Character['id'], newData: Character.UpdateRequest): Promise<void> {
+  async update(memberId: Member['id'], id: Character['id'], newData: Character.UpdateRequest): Promise<Array<string>> {
+    const changedInfo: Array<string> = [];
+
     const origin = await this.get(id);
     const date = DateTimeUtil.now();
 
@@ -489,50 +497,72 @@ export class CharactersService {
       /**
        * 0. 캐릭터 공개 여부 수정
        */
-      if (origin.isPublic !== newData.isPublic) {
+      const isVisible = origin.isPublic !== newData.isPublic;
+      if (isVisible) {
         await tx.character.update({ data: { is_public: newData.isPublic }, where: { id: id } });
       }
 
       /**
        * 1. 성격-캐릭터 관계를 업데이트 한다.
        */
-      await this.personalitiesService.updateAndDeleteMany(tx, id, origin.personalities, newData.personalities, date);
+      const isPersonalitiesChanged = this.prisma.isChanged<Personality>(origin.personalities, newData.personalities);
+
+      if (isPersonalitiesChanged) {
+        await this.personalitiesService.upsertAndDeleteMany(tx, id, origin.personalities, newData.personalities, date);
+      }
 
       /**
        * 2. 첨부파일-캐릭터 관계를 업데이트 한다.
        */
-      await this.sourcesService.deleteMany(tx, id, origin.sources, sources, date);
+      const isSourceChanged = this.prisma.isChanged<Source>(origin.sources, sources);
+
+      if (isSourceChanged) {
+        await this.sourcesService.deleteMany(tx, id, origin.sources, sources, date);
+      }
 
       /**
-       * 3. 캐릭터 스냅샷을 업데이트한다.
-       * 변경점이 없다면 마지막 스냅샷 정보를 가져온다.
+       * 3. 캐릭터 스냅샷 변경 점 여부를 조회한다.
        */
-      const isSnapshotChaged = origin.nickname !== newData.nickname || origin.image !== newData.image;
-
-      const snapshot = isSnapshotChaged
-        ? await this.createNewSnapshot(tx, {
-            characterId: id,
-            nickname: newData.nickname,
-            image: newData.image,
-            createdAt: date,
-          })
-        : await this.getLastSnapshot(id, tx);
+      const isSnapshotChanged = origin.nickname !== newData.nickname || origin.image !== (newData.image ?? null);
+      const isExperiencesChanged = this.prisma.isChanged<Experience>(origin.experiences, newData.experiences);
+      const isPositionsChanged = this.prisma.isChanged<Position>(origin.positions, positions);
+      const isSkillsChanged = this.prisma.isChanged<Skill>(origin.skills, skills);
 
       /**
-       * 4. 경력-캐릭터 스냅샷 관계를 업데이트한다.
+       * 4. 변경점이 있다면 새로운 스냅샷을 생성하고 모든 관계를 새로운 스냅샷으로 갱신한다.
+       * (변경점 : 스냅샷, 경력, 직군, 스킬명, 첨부파일이 수정되었을 ㄴ때)
        */
-      await this.experiencesService.updateAndDeleteMany(tx, snapshot.id, origin.experiences, newData.experiences, date);
+      const isChanged =
+        isSnapshotChanged || isExperiencesChanged || isPositionsChanged || isSkillsChanged || isSourceChanged;
 
-      /**
-       * 5. 직종-캐릭터 스냅샷 관계를 업데이트한다.
-       */
-      await this.positionsService.updateAndDeleteMany(tx, snapshot.id, origin.positions, positions);
+      if (isChanged) {
+        const newSnapshot = await this.createNewSnapshot(tx, {
+          characterId: id,
+          nickname: newData.nickname,
+          image: newData.image,
+          createdAt: date,
+        });
 
-      /**
-       * 6. 기술스택-캐릭터 스냅샷 관계를 업데이트한다.
-       */
-      await this.skillsService.updateAndDeleteMany(tx, snapshot.id, origin.skills, skills);
+        // 직종-캐릭터 스냅샷 관계를 업데이트한다.
+        await this.experiencesService.updateSnapshotMany(tx, newSnapshot.id, newData.experiences, date);
+
+        // 성격-캐릭터 스냅샷 관계를 업데이트한다.
+        await this.positionsService.updateSnapshotMany(tx, newSnapshot.id, positions);
+
+        // 기술 스택-캐릭터 스냅샷 관계를 업데이트한다.
+        await this.skillsService.updateSnapshotMany(tx, newSnapshot.id, skills);
+      }
+
+      isVisible ? changedInfo.push('공개 여부') : null;
+      isPersonalitiesChanged ? changedInfo.push('캐릭터 성격') : null;
+      isSourceChanged ? changedInfo.push('첨부파일') : null;
+      isSnapshotChanged ? changedInfo.push('캐릭터 정보') : null;
+      isExperiencesChanged ? changedInfo.push('경력') : null;
+      isPositionsChanged ? changedInfo.push('직종') : null;
+      isSkillsChanged ? changedInfo.push('스킬') : null;
     });
+
+    return changedInfo;
   }
 
   /**
